pyproject.toml
[project]
name = "pomodoro-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "artteri",email = "mikhail.kazakov.qa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "pomodoro_poetry", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
